requirements.txt
groq
phidata
duckduckgo-search
python-dotenv
yfinance
fastapi
uvicorn
packaging
